Fix some updater issues.
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@6168 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/sha1.c
@@ -23,7 +23,7 @@ This file came to FTE via EzQuake.
 typedef struct
 {
     unsigned int state[5];
-    size_t count[2];
+    unsigned int count[2];
     unsigned char buffer[64];
 } SHA1_CTX;
 #define SHA1_DIGEST_SIZE 20
